+++ inc/fit-db.php
@@ -731,18 +731,14 @@ function commit_loadout(&$fit, $ownerid, $accountid, &$error = null) {
 	/* Assume commit_loadout() was successful, do the post-commit
 	 * stuff */
 
-	if($fit['metadata']['visibility'] == VISIBILITY_PRIVATE) {
-		\Osmium\Search\unindex($loadoutid);
-	} else {
-		\Osmium\Search\index(
-			\Osmium\Db\fetch_assoc(
-				\Osmium\Search\query_select_searchdata(
-					'WHERE loadoutid = $1',
-					array($loadoutid)
-				)
+	\Osmium\Search\index(
+		\Osmium\Db\fetch_assoc(
+			\Osmium\Search\query_select_searchdata(
+				'WHERE loadoutid = $1',
+				array($loadoutid)
 			)
-		);
-	}
+		)
+	);
 
 	$revision = $fit['metadata']['revision'];